libft_1

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: eestell <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/13 13:18:17 by eestell           #+#    #+#             */
-/*   Updated: 2019/09/13 13:23:27 by eestell          ###   ########.fr       */
+/*   Updated: 2019/09/21 17:49:54 by eestell          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t		ft_strlen(const char *s)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (s[i] != '\0')
